Highlight matching characters
This change causes the matching characters in each choice (except for
the currently highlighted choice) to be highlighted in red. Scores have
to be recomputed over the displayed matches in order to obtain richer
match data sufficient to render the choices in this way, but since there
is a very small hard limit on the maximum number of choices displayed at
any given time, this doesn't concern me.