[Unix] Improve SIGWINCH handling
This commit changes visible_choices from a field to a method that always
recomputes the number of choices to display based on the current
dimensions of the terminal. It also introduces new screen blanking
behavior in response to SIGWINCH: instead of just clearing the lines
we've been writing on, we emit a terminal control code that blanks the
entire screen. This ensures that garbage is not left on the screen after
resizing.