Refuse to run on Cygwin
Heatseeker attempts to use the Windows console API when running in
Cygwin-type environments, which results in the process hanging. This
change causes the command to fail fast when Cygwin is detected.

It should be noted that detecting Cygwin is inordinately difficult;
there doesn't appear to be a good, stable, documented way of doing it.
The approach we take is to call Windows to obtain the file name
associated with the standard input handle, as on Cygwin this name has a
characteristic prefix or infix (on Windows it seems to be blank.) See
also https://github.com/BurntSushi/ripgrep/issues/94, where this
technique was originally developed.

Fixes #25.