Windows: Enable VT input
This change sets `ENABLE_VIRTUAL_TERMINAL_INPUT` on `CONIN`, which
enables input to be processed with the same code (`translate_bytes`,
moved here to the `ansi` module) used on the other operating systems,
with some additional preprocessing to handle UTF-16 to UTF-8 conversion,
`WINDOW_BUFFER_SIZE_EVENT`, and key-up events. (For some reason, stray
key-up events seem to come in at startup, and they need to be filtered
out.)

In addition to simplifying the code, this change improves support for
MSYS2, which seems to use virtual terminal input natively.